refactor: Improvements to WAL Manager
This commit contains a number of improvements to the WAL Manager
including better comments, moving firstAppend out of item and
into Manager, resetting SegmentWriter before acquiring the mutex,
moving logic out of NextPending and into the moveIfExceededMaxAge
function, and using channels to broadcast when there are segments
in the pending list instead of polling with FlushCheckPeriod.

All these improvements have reduced jitter and lower latency for
Push requests, and allowed us to remove FlushCheckPeriod from
the code.